Add: 1. Get current location, upload selfie 2. Local authentication. 3. Clock in and clock out service. 4. Attendance list based on selected month. 5. Check list screen image upload, 6. Signature, 7. Add rating screen. 8. Add rating in check list. 9. Model class for check list. 10. Attendance status model class. 11. History model class. 12. Job detail model class. 13. Local auth, upload location, upload image for start attendance. 14. Local auth, upload location , upload image for end attendance. 15. Today working services list in attendance screen. 16. Show selected images from check lsit screen to report screen. 17. Service detail screen design and functionalities. 18. Choose photo from gallary and camera design and functionalities.
Mod:
1. Check list design issue.
2. Detail button in shedule list screen.

+++ lib/model/attendece_status_model.dart
@@ -57,8 +57,8 @@ class MAttendanceStatusItem {
   String? siteId;
   int? serviceId;
   dynamic siteName;
-  int? latIn;
-  int? longIn;
+  double? latIn;
+  double? longIn;
   dynamic latOut;
   dynamic longOut;
   dynamic imageIn;
